+++ opentracing_instrumentation/utils.py
@@ -1,0 +1,41 @@
+# Copyright (c) 2015 Uber Technologies, Inc.
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+from __future__ import absolute_import
+
+import opentracing
+
+
+def start_child_span(operation_name, tracer=None, parent=None, tags=None):
+    """
+    Start a new span as a child of parent_span. If parent_span is None,
+    start a new root span.
+
+    :param operation_name: operation name
+    :param tracer: Tracer or None (defaults to opentracing.tracer)
+    :param parent: parent Span or None
+    :param tags: optional tags
+    :return: new span
+    """
+    tracer = tracer or opentracing.tracer
+    return tracer.start_span(
+        operation_name=operation_name,
+        child_of=parent.context if parent else None,
+        tags=tags
+    )